mcp-yandex-search/pyproject.toml
[project]
name = "mcp-yandex-search"
version = "0.1.0"
description = "MCP server for web search using Yandex Search API"
authors = [
    {name = "Yandex Search MCP Team", email = "yandex-search@example.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=0.2.0",
    "httpx>=0.27.0",
    "uvicorn>=0.30.0",
    "starlette>=0.40.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."] 